+++ Server/box/box-api/src/main/java/com/wecon/box/api/RealHisCfgDataApi.java
@@ -59,7 +59,7 @@ public interface RealHisCfgDataApi {
 	 * @param ids
 	 * @return
      */
-	boolean batchDeleteById(final List<Integer> ids);
+	boolean batchDeleteById(final List<Long> ids);
 
 	/**
 	 * 获取某条历史（实时）数据的分页列表
